+++ CHANGES.rst
@@ -17,6 +17,7 @@ Changelog
 - added `idc-layer-segments` tool to generate layer-segments output from image segmentation input files that require
   processing in terms of file name alignment and base image augmentation
 - the `is-to-od` filter now initializes the labels member variable correctly when only supplying a regexp for the labels
+- added the `discard-positives` filter that skips data that has annotations
 
 
 0.1.0 (2025-10-31)